Optimize Dockerfile and add ccache
The provided Dockerfiles contained a few items which
violate the best practices (https://docs.docker.com/engine/userguide/eng-image/dockerfile_best-practices/).
In addition to fixing these issues, the size was
reduced by removing unnecessary caches and files.
Ccache was added into the image to enable faster builds
(especially for the travis builds). To store the ccache cache
outside the container the docker run command need to be extended with

-v /tmp/ccache:/root/.ccache/

Resolves #13

Signed-off-by: Ralf Anton Beier <ralf_beier@me.com>