Added first unit test for waypoint_updater
Created a simple unit tests which provides the current pose
and base_waypoints and receives the final_waypoints.
Fixed a issue in the waypoint_updater which could lead to
an index out of bound error.

Minor cleanups in the tl_detector unit tests.

Added instructions on how to call the unit tests.

Signed-off-by: Ralf Anton Beier <ralf_beier@me.com>