Styx: workaround for socket.timeout and faster shutdown
A wrong implementation in eventlet let the server die
after some time if there was no initial connection.
It was not solved yet (https://github.com/eventlet/eventlet/issues/222).
Retrigger the server after this exception.

Ros and eventlet's are not playing too nice together
when it is time to shutdown the service.
Trigger a sigkill to enforce this.

Few pylint cleanups.

Signed-off-by: Ralf Anton Beier <ralf_beier@me.com>