If a client disconnects, set dbw_enable back to false.
If a client disconnects, reset the dbw_enable to manual mode.
This ensures that the dbw_node will also stop predicting the
next moves.

Unfortunately the simulator needs to be closed to trigger a disconnect.
Just leaving the simulation mode will keep the connection open and
when entering the simulation mode again, a new one is created.

Signed-off-by: Ralf Anton Beier <ralf_beier@me.com>