Computation of steer, throttle and brake.
Controller now computes all three values as follows:

- Steer: Composed by correcive steering and predictive steering.
We use a PID controller to compute corrective steering, for this
we use the CTE at each iteration. Predictive steering is computed
using YawController.

- Throttle and Brake: These two values are computed using a PID
controller that is based on velocity error. Depending on acceleration
given by this PID, the throttle is given ( acceleration is positive ),
or brake is given ( acceleration is negative ).