Introducing twiddle to optimize PID controller parameters.
This class implements the twiddle algorithm to optimize parameters
for a PID controller. It encapsulates a PID controller that mutates
everytime a new set of parameters that yield smaller error are found.

Once it has found the parameters that satisfy the constraints given
(mainly the tolerance) it stops experimenting.

You can use this class instead of a regular PID controller.

Following commits will use this class.