Check status of unit test based on files and not exit status
The behavior of catkin_make run_tests differs locally
from the travis environment. While on travis it always
exits with 1 a successful local build exists with 0.
To fix this the passed/failed status will be returned
by a comparition of the unit test files. If all files
do contain no errors, the test will pass otherwise it will
fail.

With the next iteration the script has to be seperated
into an own file.

Signed-off-by: Ralf Anton Beier <ralf_beier@me.com>